+++ Hit.cpp
@@ -5,7 +5,7 @@
 #include "Hit.hpp"
 
 namespace nagato {
-    Hit::Hit(double d, Vector3 p, Vector3 n, Object *sphere)
+    Hit::Hit(float d, Vector3 p, Vector3 n, Object *sphere)
             :
             distance(d), point(p), normal(n), sphere(sphere) {
