+++ Matrix4.hpp
@@ -10,14 +10,14 @@
 namespace nagato {
     class Matrix4 {
     public:
-        double data[4][4] = {0};
+        float data[4][4] = {0};
 
         Matrix4()
         = default;
 
-        explicit Matrix4(double a);
+        explicit Matrix4(float a);
 
-        explicit Matrix4(double a[4][4]);
+        explicit Matrix4(float a[4][4]);
 
         friend inline Matrix4 operator+(Matrix4 a, Matrix4 b) {
           Matrix4 m;
@@ -63,7 +63,7 @@ namespace nagato {
           return m;
         }
 
-        inline static Matrix4 transform(double x, double y, double z) {
+        inline static Matrix4 transform(float x, float y, float z) {
           Matrix4 m = identity();
           m.data[0][3] = x;
           m.data[1][3] = y;
@@ -71,27 +71,27 @@ namespace nagato {
           return m;
         }
 
-        inline static Matrix4 rotate(int a, double theta) {
+        inline static Matrix4 rotate(int a, float theta) {
           Matrix4 m = identity();
 
           if (a == 0) {
             // x軸を中心に回転
-            m.data[1][1] = cos(theta);
-            m.data[1][2] = sin(theta);
-            m.data[2][1] = -sin(theta);
-            m.data[2][2] = cos(theta);
+            m.data[1][1] = static_cast<float>(cos(theta));
+            m.data[1][2] = static_cast<float>(sin(theta));
+            m.data[2][1] = static_cast<float>(-sin(theta));
+            m.data[2][2] = static_cast<float>(cos(theta));
           } else if (a == 1) {
             // y軸を中心に回転
-            m.data[0][0] = cos(theta);
-            m.data[0][2] = -sin(theta);
-            m.data[2][0] = sin(theta);
-            m.data[2][2] = cos(theta);
+            m.data[0][0] = static_cast<float>(cos(theta));
+            m.data[0][2] = static_cast<float>(-sin(theta));
+            m.data[2][0] = static_cast<float>(sin(theta));
+            m.data[2][2] = static_cast<float>(cos(theta));
           } else if (a == 2) {
             // z軸を中心に回転
-            m.data[0][0] = cos(theta);
-            m.data[0][1] = sin(theta);
-            m.data[1][0] = -sin(theta);
-            m.data[1][1] = cos(theta);
+            m.data[0][0] = static_cast<float>(cos(theta));
+            m.data[0][1] = static_cast<float>(sin(theta));
+            m.data[1][0] = static_cast<float>(-sin(theta));
+            m.data[1][1] = static_cast<float>(cos(theta));
           }
           return m;
         }
