+++ Vector3.cpp
@@ -9,13 +9,13 @@
 
 namespace nagato {
 
-    Vector3::Vector3(double v)
+    Vector3::Vector3(float v)
             : Vector3(v, v, v) {}
 
-    Vector3::Vector3(double x, double y, double z)
+    Vector3::Vector3(float x, float y, float z)
             : x(x), y(y), z(z) {}
 
-    double Vector3::operator[](int i) {
+    float Vector3::operator[](int i) {
       return (&x)[i];
     }
